pam_ibmacf Reduce journal messages (#19)
This improves the pam_ibmacf module in two related ways.

1. This improves the set of pam_syslog messages sent by the pam_ibmacf
module.  It reduces the number of messages sent in the normal cases
and sends usable diagnostic messages.

2. Normally the ACF is checked against the primary production key, the
backup production key, and if field mode is disabled, against the
development key. However, there is no point in continuing if the first
check failed in a way that the subsequent checks would also fail. The
code was changed to give up in this case.

Tested: A variety of authentication outcomes
- ACF missing
- Success
- Bad password
- Missing keys

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>