add second backup production key
This adds a second backup production key, so there are three production
keys total.  The key is used in the pam_ibmacf authenticaton flow.

Tested:
Ensured the key is installed as intended.
Ensured authentication continues to work.
Did not test uploading the new key with respect to uploading an ACF or
authenticating a login.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>