Add "production" ACF creation command to celogin_cli
The existing `celogin_cli create` command was created as a lab/testing
only utility. As such it does not provide as robust of an interface.

This commit creates a new command `celogin_cli create_prod` that can be
used for production ACFs. The process is derived from the "out-of-band"
method that was introduced to the "create" command (which is removed in
this commit). create_prod does not implement a signature routine, since
it is assumed that the private key is not readily avaible on disk.
Instead the ACF generation is broken up into two steps, creation of the
json and then packaging of the json signature and json into the asn1
binary. The caller can then easily take the digest from the first step,
pass it into their crypto hardware/libary and then pass the resulting
signature back into the celogin_cli utility.