tacf: Support for recovering the admin account
Added admin recovery support to the admin-reset ACF
support. In addition to resetting the password the
account will also be enabled and unlocked.
Support was also added to fix up the account if it
is in a bad state due to a failed delete attempt or
other activity. If the account does not exist at all
then the account will be created.

Signed-off-by: Ben Tyner <ben.tyner@ibm.com>