Improve ACF install error reporting
This fixes an issue where the authRc variable was shadowed in the ACF install
path so it was not set to any meaningful value.

This improves the error code produced by the ACF install path.  The main
observation is that multiple public keys are attempted and all but one of
them will fail with CeLoginRc::RcBase::SignatureNotValid.  That is a valid
error code, but errors from keys where the signature was valid should take
precedence.  For example, consider when the second key is valid but the ACF
is expired: the SignatureNotValid from the third key should not mask this
more significant error code.

Tested: Yes
Scenarios: valid ACF installed, incorrect signature, expired ACF.
The caller writes the error code into a journal entry.

Signed-off-by: Joseph Reynolds <joseph.reynolds1@ibm.com>